Move api client and properties class to api directory
