+++ src/api/indicator-property.test.ts
