+++ src/custom-functions/bcode-quarter.test.ts
@@ -1,12 +1,15 @@
+import { BuffettCodeApiClientV2 as BuffettCodeApiClientV2Mock } from '../__mocks__/api/client'
 import { QuarterCache as QuarterCacheMock } from '../__mocks__/services/quarter-cache'
 
-jest.mock('../client')
+jest.mock('../api/client', () => ({
+  BuffettCodeApiClientV2: BuffettCodeApiClientV2Mock
+}))
 jest.mock('../services/quarter-cache', () => ({
   QuarterCache: QuarterCacheMock
 }))
 
 import { bcodeQuarter } from './bcode-quarter'
-import { BuffettCodeApiClientV2 } from '../client'
+import { BuffettCodeApiClientV2 } from '../api/client'
 import { QuarterCache } from '../services/quarter-cache'
 import { Result } from '../result'
 import { YearQuarter } from '../year-quarter'
