+++ src/custom-functions/bcode-quarter.ts
@@ -1,7 +1,7 @@
 import { ApiResponseError } from './error'
-import { BuffettCodeApiClientV2 } from '../client'
+import { BuffettCodeApiClientV2 } from '../api/client'
 import { QuarterCache } from '../services/quarter-cache'
-import { QuarterProperty } from '../quarter-property'
+import { QuarterProperty } from '../api/quarter-property'
 import { Result } from '../result'
 import { YearQuarterRange } from '../year-quarter-range'
 import { YearQuarter } from '../year-quarter'
