+++ src/services/csv-exporter.ts
@@ -1,9 +1,9 @@
-import { BuffettCodeApiClientV2 } from '../client'
+import { BuffettCodeApiClientV2 } from '../api/client'
 import { Setting } from '../setting'
 import { YearQuarter } from '../year-quarter'
 import { YearQuarterRange } from '../year-quarter-range'
 import { QuarterCache } from './quarter-cache'
-import { QuarterProperty } from '../quarter-property'
+import { QuarterProperty } from '../api/quarter-property'
 
 export class CsvExporter {
   private constructor() {
